+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# 0.5.2 (11-10-2025)
+# 0.6.0 (11-10-2025)
+
+- **breaking:** Replace `debounceDelay` with `updateMode` and `updateDelay` in `useRemark` hook and `<Remark>` component.
+
+## 0.5.2 (11-10-2025)
 
 - **fix:** Ensure useDebounce follows React rules of hooks
 - **fix:** Debounce key instead of text in useSpeechInternal to prevent starvation
@@ -6,7 +10,7 @@
 
 ## 0.5.0 (11-10-2025)
 
-- **feat:** Add `debounceDelay` option to `useRemark` hook to debounce changes to the `markdown` input, reducing redundant processing during rapid updates.
+- **feat:** Add `debounceDelay` option to `useRemark` hook and `<Remark>` component to debounce changes to the `markdown` input, reducing redundant processing during rapid updates.
 
 ## 0.4.0 (01-07-2025)